+++ eslint.config.mjs
@@ -9,6 +9,7 @@ import js from '@eslint/js';
 import lit from 'eslint-plugin-lit';
 import litA11Y from 'eslint-plugin-lit-a11y';
 import path from 'node:path';
+import sds from '@solid-design-system/eslint-plugin';
 import sortImportsEs6Autofix from 'eslint-plugin-sort-imports-es6-autofix';
 import tsParser from '@typescript-eslint/parser';
 import typescriptEslint from '@typescript-eslint/eslint-plugin';
@@ -46,6 +47,12 @@ export default [
       'packages/theming/types/'
     ]
   },
+
+  {
+    ...sds.configs.recommended,
+    files: ['**/*.{html,jsx}']
+  },
+
   ...compat.extends(
     'eslint:recommended',
     'plugin:wc/recommended',