+++ packages/components/src/components/input/input.test.ts
@@ -61,6 +61,13 @@ describe('<sd-input>', () => {
     expect(input.disabled).to.be.true;
   });
 
+  it('should have aria-disabled when visually-disabled', async () => {
+    const el = await fixture<SdInput>(html` <sd-input visually-disabled></sd-input> `);
+    const input = el.shadowRoot!.querySelector<HTMLInputElement>('[part~="input"]')!;
+
+    expect(input.getAttribute('aria-disabled')).to.equal('true');
+  });
+
   it('should render label wrapper if label or tooltip are present', async () => {
     const el = await fixture<SdInput>(html`
       <sd-input label="test">