+++ packages/components/src/components/radio-button/radio-button.test.ts
@@ -21,6 +21,35 @@ describe('<sd-radio-button>', () => {
     expect(radio2.checked).to.be.false;
   });
 
+  it('should not become checked when clicked if visually-disabled', async () => {
+    const radioGroup = await fixture<SdRadioGroup>(html`
+      <sd-radio-group value="1">
+        <sd-radio-button id="radio-1" value="1" visually-disabled></sd-radio-button>
+        <sd-radio-button id="radio-2" value="2" visually-disabled></sd-radio-button>
+      </sd-radio-group>
+    `);
+    const radio1 = radioGroup.querySelector<SdRadioButton>('#radio-1')!;
+    const radio2 = radioGroup.querySelector<SdRadioButton>('#radio-2')!;
+    const radio2Button = radio2.shadowRoot!.querySelector('button');
+
+    radio2.click();
+    await Promise.all([radio1.updateComplete, radio2.updateComplete]);
+
+    expect(radio2Button!.getAttribute('aria-checked')).to.equal('true');
+  });
+
+  it('should have aria-disabled when visually-disabled', async () => {
+    const radioGroup = await fixture<SdRadioGroup>(html`
+      <sd-radio-group value="">
+        <sd-radio-button id="radio-1" value="1" visually-disabled></sd-radio-button>
+      </sd-radio-group>
+    `);
+
+    const radio1 = radioGroup.querySelector<SdRadioButton>('#radio-1')!;
+    const radio1Button = radio1.shadowRoot!.querySelector('button');
+    expect(radio1Button!.getAttribute('aria-disabled')).to.equal('true');
+  });
+
   it('should receive positional classes from <sd-button-group>', async () => {
     const radioGroup = await fixture<SdRadioGroup>(html`
       <sd-radio-group value="1">