+++ packages/docs/src/stories/components/radio-button.stories.ts
@@ -207,3 +207,40 @@ export const Disabled = {
     </div>
   `
 };
+
+/**
+ * Use the `visually-disabled` attribute to style the component as if it was disabled and enable `aria-disabled` to allow it to be reachable by screen readers.
+ *
+ * __Hint:__ When using this attribute, make sure to provide ways to inform the user why the element is disabled and how to enable it. This can be done by using the `help-text` attribute or wrapping the element in a sd-tooltip.
+ *
+ * **Accessibility Hint:** Disabling elements is not recommended for accessibility reasons.
+ */
+export const VisuallyDisabled = {
+  name: 'Visually Disabled',
+  render: () => html`
+    <div class="flex gap-12 h-[100px] pt-12">
+      <sd-radio-group>
+        <sd-tooltip content="Visually Disabled" trigger="click focus" size="sm">
+          <sd-radio-button value="1" visually-disabled>
+            <sd-icon name="system/image" slot="icon"></sd-icon>
+            Label
+          </sd-radio-button>
+        </sd-tooltip>
+
+        <sd-tooltip content="Visually Disabled" trigger="click focus" size="sm">
+          <sd-radio-button value="2" visually-disabled>
+            <sd-icon name="system/image" slot="icon"></sd-icon>
+            Disabled
+          </sd-radio-button>
+        </sd-tooltip>
+
+        <sd-tooltip content="Visually Disabled" trigger="click focus" size="sm">
+          <sd-radio-button value="3" visually-disabled>
+            <sd-icon name="system/image" slot="icon"></sd-icon>
+            Label
+          </sd-radio-button>
+        </sd-tooltip>
+      </sd-radio-group>
+    </div>
+  `
+};