+++ packages/docs/src/stories/components/radio-group.stories.ts
@@ -15,7 +15,7 @@ const { overrideArgs } = storybookHelpers('sd-radio-group');
  *
  * **Related templates**:
  * - [Radio Button Group](?path=/docs/templates-radio-button-group--docs)
- * - [Tooltip](?path=/docs/templates-tooltip--docs)
+ * - [Radio Group with Tooltip](?path=/docs/templates-tooltip--docs)
  */
 
 export default {
@@ -30,11 +30,6 @@ export default {
     }
   },
   args: overrideArgs([
-    {
-      type: 'slot',
-      name: 'label',
-      value: `<label slot="label">Group Label</label>`
-    },
     {
       type: 'slot',
       name: 'default',
@@ -99,6 +94,25 @@ export const Label = {
       <sd-radio value="2">Radio 2</sd-radio>
       <sd-radio value="3">Radio 3</sd-radio>
     </sd-radio-group>
+  `
+};
+
+/**
+ * Use the `help-text` attribute to add a descriptive “help text”.
+ *
+ * For help texts that contain HTML, use the `help-text` slot instead.
+ */
+export const HelpText = {
+  name: 'Help Text',
+  render: () => html`
+    <div class="flex gap-12">
+      <sd-radio-group orientation="vertical" help-text="Help text attribute">
+        <label slot="label">Group Label</label>
+        <sd-radio value="1">Radio 1</sd-radio>
+        <sd-radio value="2">Radio 2</sd-radio>
+        <sd-radio value="3">Radio 3</sd-radio>
+      </sd-radio-group>
+    </div>
   `
 };